Build warning

+++ src/libs/svg/lv_svg_decoder.c
@@ -317,7 +317,7 @@ static uint8_t * alloc_file(const char * filename, uint32_t * size)
     /*Read file to buffer*/
     data = lv_malloc(data_size);
     if(data == NULL) {
-        LV_LOG_WARN("malloc failed for data size %lu", data_size);
+        LV_LOG_WARN("malloc failed for data size %lu", (long unsigned)data_size);
         goto failed;
     }
